Added the Micro-Manager Video4Linux device adapter
This adds the Video4Linux device adapter that is packaged with Micro-Manager to the list of
Raspberry Pi-supported device adapters. While this device adapter is very limited in functionality,
it does provide support for multiple cameras, including the Raspberry Pi camera module.

The mm.py script is also updated to include device file detection, which is needed when multiple
cameras are installed on a Pi.